Better use double precision

+++ src/lib/renderer/enesim_renderer_raddist.h
@@ -34,8 +34,8 @@ EAPI double enesim_renderer_raddist_factor_get(Enesim_Renderer *r);
 EAPI void enesim_renderer_raddist_source_surface_set(Enesim_Renderer *r, Enesim_Surface *src);
 EAPI Enesim_Surface * enesim_renderer_raddist_source_surface_get(Enesim_Renderer *r);
 
-EAPI void enesim_renderer_raddist_x_set(Enesim_Renderer *r, int ox);
-EAPI void enesim_renderer_raddist_y_set(Enesim_Renderer *r, int oy);
+EAPI void enesim_renderer_raddist_x_set(Enesim_Renderer *r, double ox);
+EAPI void enesim_renderer_raddist_y_set(Enesim_Renderer *r, double oy);
 
 /**
  * @}
